agentaid-voice/pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "LiveKit AI Agent"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "livekit",
    "livekit-agents",
    "livekit-plugins-noise-cancellation",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector",
    "torch",
    "transformers",
]
